+++ src/Credfeto.Version.Information.Generator/Models/ErrorInfo.cs
@@ -4,16 +4,19 @@ using Microsoft.CodeAnalysis;
 
 namespace Credfeto.Version.Information.Generator.Models;
 
-[DebuggerDisplay("{Location} {Exception}")]
+[DebuggerDisplay("{LocationString} {ExceptionMessage}")]
 public readonly record struct ErrorInfo
 {
     public ErrorInfo(Location location, Exception exception)
     {
-        this.Location = location;
-        this.Exception = exception;
+        this.LocationString = location.ToString();
+        this.ExceptionMessage = exception.Message;
+        this.ExceptionStackTrace = exception.StackTrace;
     }
 
-    public Location Location { get; }
+    public string LocationString { get; }
 
-    public Exception Exception { get; }
+    public string ExceptionMessage { get; }
+
+    public string? ExceptionStackTrace { get; }
 }